\name{MEC}

\alias{MEC}
\alias{MEC-package}

\docType{package}

\title{MEC}

\description{
The MEC package is a collection of algorithms used during the Math+Econ+Code class.
}

\details{
\tabular{ll}{
Package: \tab MEC\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2018-05-22\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab Yes\cr
Depends: \tab Rcpp, RcppArmadillo \cr }
}
\section{How to cite this package}{
 \preformatted{@Manual{MEC2018,
 author       = {Keith O'Hara},
 title        = {{MEC}: Algorithms for the Math+Econ+Code class.},
 year         = {2018},
 note         = {R package version 0.1.0.}}}
\emph{}
}

\section{License}{
MEC is licensed under the GNU General Public License (GPL) version 2, or (at your option) any later version.
}

\author{Keith O'Hara}